Section edit bugfix
When you would update a section, before this fix it would redirect you to the links page and bring you to the ID of whatever link that was, sometimes that link would not exist. That has been fixed, and now when you edit a section it will  bring you back to the section.